Migrate isolated functionality to better homes (#1085)
* Migrate GaloisRingSharedIris to protocol submodule

* Migrate random individual share generation to HNSW bench executable

* Revise module path and a function name for GaloisRingSharedIris

* Update local iris code sharing to return fixed length array

---------
