+++ iris-mpc-cpu/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod database_generators;
 pub mod execution;
 pub mod hawkers;
 pub mod hnsw;
@@ -7,4 +6,4 @@ pub(crate) mod network;
 pub(crate) mod proto_generated;
 pub mod protocol;
 pub mod py_bindings;
-pub(crate) mod shares;
+pub mod shares;
