+++ iris-mpc-cpu/src/protocol/ops.rs
@@ -1,11 +1,11 @@
 use super::binary::{mul_lift_2k, single_extract_msb_u32};
 use crate::{
-    database_generators::GaloisRingSharedIris,
     execution::session::{BootSession, Session, SessionHandles},
     network::value::NetworkValue::{self},
     protocol::{
         binary::{lift, open_bin},
         prf::{Prf, PrfSeed},
+        shared_iris::GaloisRingSharedIris,
     },
     shares::{
         bit::Bit,
@@ -277,13 +277,12 @@ pub async fn compare_threshold_and_open(
 mod tests {
     use super::*;
     use crate::{
-        database_generators::generate_galois_iris_shares,
         execution::{
             local::{generate_local_identities, LocalRuntime},
             player::Identity,
         },
         hawkers::plaintext_store::PlaintextIris,
-        protocol::ops::NetworkValue::RingElement32,
+        protocol::{ops::NetworkValue::RingElement32, shared_iris::GaloisRingSharedIris},
         shares::{int_ring::IntRing2k, ring_impl::RingElement},
     };
     use aes_prng::AesRng;
@@ -556,8 +555,10 @@ mod tests {
 
         let iris_db = IrisDB::new_random_rng(2, &mut rng).db;
 
-        let first_entry = generate_galois_iris_shares(&mut rng, iris_db[0].clone());
-        let second_entry = generate_galois_iris_shares(&mut rng, iris_db[1].clone());
+        let first_entry =
+            GaloisRingSharedIris::generate_shares_locally(&mut rng, iris_db[0].clone());
+        let second_entry =
+            GaloisRingSharedIris::generate_shares_locally(&mut rng, iris_db[1].clone());
 
         let mut jobs = JoinSet::new();
         for (index, player) in runtime.get_identities().iter().cloned().enumerate() {
